Extract FBXModel to separate module
